bots/bot1/requirements.txt
ccxt
polars
websocket-client
python-dotenv
numpy
lightgbm
scikit-learn
joblib
sqlalchemy
psycopg2-binary